% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_ind.R
\name{apa_t_ind}
\alias{apa_t_ind}
\title{APA text for Independent-Samples T-Test}
\usage{
apa_t_ind(a, b, group_a = "group A", group_b = "group B", dv = "scores")
}
\arguments{
\item{a}{A vector (e.g., the values in the control condition).}

\item{b}{A vector (e.g., the values in the experimental condition).}

\item{group_a}{The text used to describe group A.}

\item{group_b}{The text used to describe group B.}
}
\value{
A character string.
}
\description{
Create APA-formatted text for the results of an independent-samples t-test in the following format:
}
\details{
Compared to the {n_A} participants in {group_a} (M = {mean_A}, SD = {sd_A}), the {n_B} participants in {group_b} (M = {mean_B}, SD = {sd_B}) had {is_sig}significantly {direction} values, t({df}) = {t_value}, p = {p_value}.
}
\examples{
ctl <- rnorm(20, 0, 1)
exp <- rnorm(30, 0.8, 1)

apa_t_ind(ctl, exp)
apa_t_ind(ctl, exp,
          "the control group",
          "the experimental group",
          "brain scores")
}
