% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{calc_monthly_spend}
\alias{calc_monthly_spend}
\title{Calculate monthly spend}
\usage{
calc_monthly_spend(transactions)
}
\arguments{
\item{transactions}{\code{data.frame} A data set with transactions indexed by customer and month}
}
\value{
\code{data.frame} A data set with aggregated monthly spend per customer
}
\description{
A more extensive description
}
