% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_xl.R
\name{count_xl}
\alias{count_xl}
\title{Count excel files}
\usage{
count_xl(type = c("xlsx", "xls", "both"), recurse = TRUE)
}
\arguments{
\item{type}{What type of excel files to look for. You can search for XLSX,
XLS or both just to be sure.}

\item{recurse}{Number of levels to recurse or TRUE to recurse fully}
}
\value{
The number of excel files
}
\description{
Let's count the number of excel files you have
}
\examples{
count_xl()
}
