% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg-package.R
\docType{package}
\name{demopkg-package}
\alias{demopkg}
\alias{demopkg-package}
\title{A Demo for Coding Club 2022.}
\description{
The goal of demopkg is to practice making a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kai-lim/demopkg}
  \item \url{https://kai-lim.github.io/demopkg/}
  \item Report bugs at \url{https://github.com/kai-lim/demopkg/issues}
}

}
\author{
\strong{Maintainer}: Kai Lim \email{kxlim94@gmail.com}

}
\keyword{internal}
