% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamonds_df}
\alias{diamonds_df}
\title{A small data frame about diamonds}
\format{
A data frame with 1000 rows and 3 variables
\describe{
\item{clarity}{The diamond's clarity. A categorical variable.}
\item{depth}{The diamond's depth. A numerical variable.}
\item{price}{The price of the diamond. A numercial variable}
}
}
\usage{
diamonds_df
}
\description{
This data set is a subset of the diamonds data set
from the 'ggplot2' package.
}
\keyword{datasets}
