% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example data}
\format{
\subsection{\code{example_dataset}}{

A data frame with 3 rows and 1 column:
\describe{
\item{x}{A number, between 1 and 3}
}
}
}
\source{
\url{https://nisra.gov.uk/geography-lookups}
}
\usage{
example_dataset
}
\description{
This could be a lookup for geographies. The source isn't real.
}
\keyword{datasets}
