% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{ilogit}
\alias{ilogit}
\title{Calculate the inverse of the logit function}
\usage{
ilogit(logodds)
}
\arguments{
\item{logodds}{The logarithm (to base e) of some probability}
}
\value{
The probability that is the inverse of the log odds
}
\description{
Calculate the inverse of the logit function
}
\examples{
ilogit(3)
ilogit(-2)
ilogit(0)
}
