% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{likelihood}
\alias{likelihood}
\title{Plot a binomial likelihood function}
\usage{
likelihood(n = 100, m = 60)
}
\arguments{
\item{n}{Number of trials}

\item{m}{Number of successes}
}
\value{
A ggplot of the likelihood function
}
\description{
Plot a binomial likelihood function
}
\examples{
likelihood(75, 25)
likelihood(100, 33)
likelihood(50, 10)
}
