% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_utils.R
\name{logit}
\alias{logit}
\title{Calculate the logarithm of the odds of a probability}
\usage{
logit(p)
}
\arguments{
\item{p}{A probability (a values > 0 and less than 1)}
}
\value{
The logarithm (to base e) of the probability
}
\description{
Calculate the logarithm of the odds of a probability
}
\examples{
logit(0.5)
logit(0.75)
logit(0.25)
}
