% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{median_first_ten}
\alias{median_first_ten}
\title{Annoyingly return the median of the first 10 elements of a vector}
\usage{
median_first_ten(x)
}
\arguments{
\item{data}{A dataframe.}
}
\description{
Annoyingly return the median of the first 10 elements of a vector
}
