% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{my_fun}
\alias{my_fun}
\title{Scale input to values between 0 and 1}
\usage{
my_fun(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric vector of scaled values
}
\description{
Scale input to values between 0 and 1
}
\examples{

my_fun(c(1, 10, 4))

}
