% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_pair.R
\name{round0}
\alias{round0}
\title{Round with Trailing Zeroes}
\usage{
round0(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places to be used}
}
\value{
character string of formatted number
}
\description{
Round with Trailing Zeroes
}
