% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_res_att.R
\docType{data}
\name{self_res_att}
\alias{self_res_att}
\title{Attractiveness of Self-Resembling Faces.}
\format{
A data frame with 108 rows and 16 variables:
\describe{
\item{id}{participant unique ID}
\item{sex}{sex of the participant (male/female)}
\item{ethgroup}{ethnic group of the participant (east_asian/west_asian/white)}
\item{age}{age of the participant in years}
\item{m_non}{mean number of times the other group members chose that male face as more attractive}
\item{f_non}{mean number of times the other group members chose that female face as more attractive}
\item{m_self}{number of times out of a possible 6 chose their male self-res face as more attractive}
\item{f_self}{number of times out of a possible 6 chose their female self-res face as more attractive}
\item{grpsize}{size of the group}
\item{group}{unique group ID}
\item{mascpref}{masculinity preference on an unrelated face preference task}
\item{obro}{number of older brothers}
\item{osis}{number of older sisters}
\item{ybro}{number of younger brothers}
\item{ysis}{number of younger sisters}
\item{birthorder}{birth order (only/firstborn/middleborn/lastborn) as calculated from number of younger and older brothers and sisters}
}
}
\source{
\url{https://osf.io/3c5s4/}
}
\usage{
self_res_att
}
\description{
Attractiveness judgements of same-sex and other-sex self-resembling faces from DeBruine (2004).
}
\keyword{datasets}
