% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squared.}
}
\value{
The square of x.
}
\description{
Takes a number x and returns its square, x * x.
}
\examples{
square(2)
square(-5.7)

}
