% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum_two_numbers}
\alias{sum_two_numbers}
\title{Summing Two Numbers}
\usage{
sum_two_numbers(number1, number2)
}
\arguments{
\item{number1}{first summand.}

\item{number2}{second summand.}
}
\value{
The sum of \code{number1} and \code{number2}.
}
\description{
Summing Two Numbers
}
\examples{
sum_two_numbers(4, 1)
}
