% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_pair.R
\name{apa_t_pair}
\alias{apa_t_pair}
\title{APA Text for Paired Samples T-Test}
\usage{
apa_t_pair(
  x,
  y,
  dv = "the DV",
  level1 = "level 1",
  level2 = "level 2",
  alpha = 0.05,
  conf.level = 0.95
)
}
\arguments{
\item{x}{A (non-empty) numeric vector of data values for level 1.}

\item{y}{A (non-empty) numeric vector of data values for level 2.}

\item{dv}{A character string describing the DV in the output text.}

\item{level1}{A character string describing level 1 in the output text.}

\item{level2}{A character string describing level 2 in the output text.}

\item{alpha}{The critical alpha for determining statistical significance.}

\item{conf.level}{The confidence level for specifying the coverage of confidence intervals.}
}
\value{
T character string combining an APA style t-test with a description of the results.
}
\description{
Create APA formatted text for the results of a paired samples t-test in the following format:

A paired-samples t-test was conducted to compare \{dv\} between \{level1\} (M = \{mean1\}, SD = \{sd1\}) and \{level2\} (M = \{mean2\}, SD = \{sd2\}). Using an alpha of \{alpha\}, there was a \{non\}significant difference; t(\{df\}) = \{t_value\}, \{p_value\}, mean difference = \{mean_diff\}, \{CI\}\% CI = [\{lower_CI\}, \{upper_CI\}].
}
\examples{
# Use generic text
apa_t_pair(x = self_res_att$f_self,
           y = self_res_att$f_non)
apa_t_pair(x = self_res_att$f_self,
           y = self_res_att$f_non,
           dv = "preferences for female faces",
           level1 = "participants who resembled those faces",
           level2 = "non-self participants")
}
