% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg-package.R
\docType{package}
\name{demopkg-package}
\alias{demopkg}
\alias{demopkg-package}
\title{A Demo for Coding Club 2022.}
\description{
The goal of demopkg is to practice making a package.
}
\author{
\strong{Maintainer}: Sam Van Horne \email{sam.vanhorne@gmail.com} (\href{https://orcid.org/0000-0002-5560-0914}{ORCID})

}
\keyword{internal}
