% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg-package.R
\docType{package}
\name{demopkg-package}
\alias{demopkg}
\alias{demopkg-package}
\title{demopkg: A Demonstration of R Packages}
\description{
This package does stuff.But it really is just a demo package, so it doesn't really do anything.
}
\author{
\strong{Maintainer}: Julia Pantin \email{Julia.Pantin@dfo-mpo.gc.ca}

}
\keyword{internal}
