% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg-package.R
\docType{package}
\name{demopkg-package}
\alias{demopkg}
\alias{demopkg-package}
\title{A Demo for Coding Club 2022.}
\description{
The goal of demopkg is to practice making a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marianmarchal/demopkg}
  \item Report bugs at \url{https://github.com/marianmarchal/demopkg/issues}
}

}
\author{
\strong{Maintainer}: Marian Marchal \email{marchal@coli.uni-saarland.de}

}
\keyword{internal}
