% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg-package.R
\docType{package}
\name{demopkg-package}
\alias{demopkg}
\alias{demopkg-package}
\title{A Demo for Coding Club 2022.}
\description{
The goal of demopkg is to practice making a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PsyTeachR/demopkg}
  \item Report bugs at \url{https://github.com/PsyTeachR/demopkg/issues}
}

}
\author{
\strong{Maintainer}: Lisa DeBruine \email{debruine@gmail.com} (\href{https://orcid.org/0000-0002-7523-5539}{ORCID})

}
\keyword{internal}
