% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopkg-package.R
\docType{package}
\name{demopkg-package}
\alias{demopkg}
\alias{demopkg-package}
\title{demopkg: A Demo Package for Demostrating R Packages}
\description{
This package does a something. It's just a demostration package though, so don't expect much. However, I would like to explore what happens with formatting. Did I just create a new paragraph. Or does this do it as well?
}
\author{
\strong{Maintainer}: Ron Lewis \email{ron.lewis@dfo-mpo.gc.ca}

}
\keyword{internal}
