% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{my_fun}
\alias{my_fun}
\title{Scales input to 0 to 1}
\usage{
my_fun(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric vector of scaled values
}
\description{
Scales input to 0 to 1
}
\examples{

my_fun(1:10)

}
