% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_pair.R
\name{round0}
\alias{round0}
\title{Round your number while retaining trailing zeroes}
\usage{
round0(x, digits = 0)
}
\arguments{
\item{x}{Vector of numbers to round.}

\item{digits}{Number of digits to round your number to.}
}
\value{
A character string for your rounded number.
}
\description{
Round your number while retaining trailing zeroes
}
