% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_pair.R
\name{round0}
\alias{round0}
\title{Control Zeros on Rounding Digits}
\usage{
round0(x, digits = 0)
}
\arguments{
\item{x}{A (non-empty) numeric vector of data values.}

\item{digits}{Integer indicating the number of decimal places (round).}
}
\value{
A character string providing the rounded value with specified number of decimal places.
}
\description{
round0 rounds the values in its first argument to the specified number of decimal places. In contrast to round, this function retains trailing zeroes to comply with APA style.
}
