% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_pair.R
\name{round0}
\alias{round0}
\title{Custom Rounding Function}
\usage{
round0(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Integer indicating the number of decimal places to be used.}
}
\value{
A character string of the formatted x value(s)
}
\description{
Rounds a number and returns a string with the required number of digits after the decimal point, including trailing zeroes, if applicable.
}
